+++ CHANGES.md
@@ -23,6 +23,7 @@ To be released.
  -  Added `BlockChain<T>.MakeTransaction(PrivateKey, IAction,
     IImmutableSet<Address>, DateTimeOffset?)` overloaded method.
     [[#2151], [#2273]]
+ -  Added `GetInnerActionTypeName()` method. [[#1910], [#2189]]
  -  (Libplanet.Explorer) Added `LibplanetExplorerSchema` class.
     [[#2065], [#2198]]
 
@@ -37,10 +38,11 @@ To be released.
 
 ### CLI tools
 
+[#1910]: https://github.com/planetarium/libplanet/issues/1910
 [#2065]: https://github.com/planetarium/libplanet/issues/2065
-[#2198]: https://github.com/planetarium/libplanet/pull/2198
 [#2273]: https://github.com/planetarium/libplanet/pull/2273
 [#2283]: https://github.com/planetarium/libplanet/pull/2283
+[#2189]: https://github.com/planetarium/libplanet/pull/2189
 [Bencodex 0.5.0]: https://www.nuget.org/packages/Bencodex/0.5.0